Remove the useless cutoff variable
This variable is not useful in R because the 0.99999999999999999999999999
is rounded to one. The substitution in the LR array is not useful at
all. Moreover, the division by 0 is not an issue in R (the value is set
to +Inf or -Inf depending upon the sign of the division). This is a
non-breaking change regarding the calibration process.